requirements.txt
SQLAlchemy
fastapi
pydantic
selenium
passlib[bcrypt]
starlette
Pillow